Reused RunningTotalIndicator
