+++ ta4j-core/src/main/java/org/ta4j/core/indicators/SMAIndicator.java
@@ -24,6 +24,7 @@
 package org.ta4j.core.indicators;
 
 import org.ta4j.core.Indicator;
+import org.ta4j.core.indicators.helpers.RunningTotalIndicator;
 import org.ta4j.core.num.Num;
 
 /**
@@ -34,12 +35,9 @@ import org.ta4j.core.num.Num;
  */
 public class SMAIndicator extends CachedIndicator<Num> {
 
-    private final Indicator<Num> indicator;
     private final int barCount;
-    private Num previousSum = zero();
+    private RunningTotalIndicator previousSum;
 
-    // serial access detection
-    private int previousIndex = -1;
 
     /**
      * Constructor.
@@ -49,52 +47,20 @@ public class SMAIndicator extends CachedIndicator<Num> {
      */
     public SMAIndicator(Indicator<Num> indicator, int barCount) {
         super(indicator);
-        this.indicator = indicator;
+        this.previousSum = new RunningTotalIndicator(indicator,barCount);
         this.barCount = barCount;
     }
 
     @Override
     protected Num calculate(int index) {
-        // serial access can benefit from previous partial sums
-        // which saves a lot of CPU work for very long barCounts
-        if (previousIndex != -1 && previousIndex == index - 1) {
-            return fastPath(index);
-        }
-
-        return slowPath(index);
-    }
-
-    private Num fastPath(final int index) {
-        var newSum = partialSum(index);
         final int realBarCount = Math.min(barCount, index + 1);
-        updatePartialSum(index, newSum);
-        return newSum.dividedBy(numOf(realBarCount));
-    }
-
-    private Num slowPath(final int index) {
-        Num sum = zero();
-        for (int i = Math.max(0, index - barCount + 1); i <= index; i++) {
-            sum = sum.plus(indicator.getValue(i));
-        }
-
-        final int realBarCount = Math.min(barCount, index + 1);
-        updatePartialSum(index, sum);
+        final var sum = partialSum(index);
         return sum.dividedBy(numOf(realBarCount));
     }
 
-    private void updatePartialSum(final int index, final Num sum) {
-        previousIndex = index;
-        previousSum = sum;
-    }
 
     private Num partialSum(int index) {
-        var sum = this.previousSum.plus(indicator.getValue(index));
-
-        if (index >= barCount) {
-            return sum.minus(indicator.getValue(index - barCount));
-        }
-
-        return sum;
+        return this.previousSum.getValue(index);
     }
 
     /** @return {@link #barCount} */
