+++ ta4j-core/src/main/java/org/ta4j/core/indicators/helpers/RunningTotalIndicator.java
@@ -36,6 +36,10 @@ import org.ta4j.core.num.Num;
 public class RunningTotalIndicator extends CachedIndicator<Num> {
     private final Indicator<Num> indicator;
     private final int barCount;
+    private Num previousSum = zero();
+
+    // serial access detection
+    private int previousIndex = -1;
 
     public RunningTotalIndicator(Indicator<Num> indicator, int barCount) {
         super(indicator);
@@ -43,22 +47,57 @@ public class RunningTotalIndicator extends CachedIndicator<Num> {
         this.barCount = barCount;
     }
 
+
     @Override
     protected Num calculate(int index) {
+        // serial access can benefit from previous partial sums
+        // which saves a lot of CPU work for very long barCounts
+        if (previousIndex != -1 && previousIndex == index - 1) {
+            return fastPath(index);
+        }
+
+        return slowPath(index);
+    }
+
+    private Num fastPath(final int index) {
+        var newSum = partialSum(index);
+        updatePartialSum(index, newSum);
+        return newSum;
+    }
+
+    private Num slowPath(final int index) {
         Num sum = zero();
-        for (int i = Math.max(getBarSeries().getBeginIndex(), index - barCount + 1); i <= index; i++) {
+        for (int i = Math.max(0, index - barCount + 1); i <= index; i++) {
             sum = sum.plus(indicator.getValue(i));
         }
+
+        updatePartialSum(index, sum);
         return sum;
     }
 
-    @Override
-    public int getUnstableBars() {
-        return 0;
+    private void updatePartialSum(final int index, final Num sum) {
+        previousIndex = index;
+        previousSum = sum;
+    }
+
+    private Num partialSum(int index) {
+        var sum = this.previousSum.plus(indicator.getValue(index));
+
+        if (index >= barCount) {
+            return sum.minus(indicator.getValue(index - barCount));
+        }
+
+        return sum;
     }
 
     @Override
     public String toString() {
         return getClass().getSimpleName() + " barCount: " + barCount;
+    }
+
+
+    @Override
+    public int getUnstableBars() {
+        return barCount;
     }
 }
